move xml and json assets to public headers
