+++ headers/3rdparty/rapidjson/ostreamwrapper.h
