+++ sources/enduro2d/high/assets/prefab_asset.cpp
@@ -6,7 +6,7 @@
 
 #include <enduro2d/high/assets/prefab_asset.hpp>
 
-#include "json_asset.hpp"
+#include <enduro2d/high/assets/json_asset.hpp>
 
 namespace
 {
