+++ sources/enduro2d/high/assets/shader_asset.cpp
@@ -6,7 +6,7 @@
 
 #include <enduro2d/high/assets/shader_asset.hpp>
 
-#include "json_asset.hpp"
+#include <enduro2d/high/assets/json_asset.hpp>
 #include <enduro2d/high/assets/text_asset.hpp>
 
 namespace
